--- a/c-deps/libroach/db.h
+++ b/c-deps/libroach/db.h
@@ -18,8 +18,23 @@
 #include <rocksdb/write_batch.h>
 #include <rocksdb/write_batch_base.h>
 
+const DBStatus kSuccess = {NULL, 0};
+
+struct DBIterator {
+  std::unique_ptr<rocksdb::Iterator> rep;
+};
+
+bool DecodeKey(rocksdb::Slice buf, rocksdb::Slice* key, int64_t* wall_time, int32_t* logical);
+
 // ToString returns a c++ string with the contents of a DBSlice.
 std::string ToString(DBSlice s);
+
+rocksdb::Slice ToSlice(DBSlice s);
+
+// ToDBString converts a slice to a data+len DBString.
+DBString ToDBString(const rocksdb::Slice& s);
+
+DBSlice ToDBSlice(const rocksdb::Slice& s);
 
 // MVCC keys are encoded as <key>[<wall_time>[<logical>]]<#timestamp-bytes>. A
 // custom RocksDB comparator (DBComparator) is used to maintain the desired
@@ -44,3 +59,13 @@ const ::rocksdb::Comparator* CockroachComparator();
 // Stats are only computed for keys between the given range.
 MVCCStatsResult MVCCComputeStatsInternal(::rocksdb::Iterator* const iter_rep, DBKey start, DBKey end,
                                          int64_t now_nanos);
+
+// DBSstFileWriterAddRaw is used internally and in ccl -- DBSstFileWriterAdd is
+// the preferred function for Go callers.
+DBStatus DBSstFileWriterAddRaw(DBSstFileWriter* fw, rocksdb::Slice key, rocksdb::Slice val);
+
+extern "C" {
+char* __attribute__((weak)) prettyPrintKey(DBKey);
+}  // extern "C"
+
+const char* DBIterAdvance(DBIterator* iter, bool skip_current_key_versions);