--- a/c-deps/libroach/include/libroachccl.h
+++ b/c-deps/libroach/include/libroachccl.h
@@ -18,6 +18,17 @@ extern "C" {
 // specified start and end keys and computes the MVCCStatsResult for it.
 DBStatus DBBatchReprVerify(DBSlice repr, DBKey start, DBKey end, int64_t now_nanos, MVCCStatsResult* stats);
 
+DBStatus DBIncIterToSst(
+  DBIterator* iter,
+  DBKey start,
+  DBKey end,
+  bool revisions,
+  DBString* data,
+  int64_t* entries,
+  int64_t* data_size,
+  DBString* intentErr
+);
+
 #ifdef __cplusplus
 }  // extern "C"
 #endif