--- a/pkg/ccl/storageccl/engineccl/mvcc.go
+++ b/pkg/ccl/storageccl/engineccl/mvcc.go
@@ -9,11 +9,17 @@
 package engineccl
 
 import (
+	"bytes"
+
+	"github.com/cockroachdb/cockroach/pkg/keys"
 	"github.com/cockroachdb/cockroach/pkg/roachpb"
 	"github.com/cockroachdb/cockroach/pkg/storage/engine"
 	"github.com/cockroachdb/cockroach/pkg/storage/engine/enginepb"
+	"github.com/cockroachdb/cockroach/pkg/util/encoding"
 	"github.com/cockroachdb/cockroach/pkg/util/hlc"
+	"github.com/cockroachdb/cockroach/pkg/util/log"
 	"github.com/pkg/errors"
+	"golang.org/x/net/context"
 )
 
 // MVCCIncrementalIterator iterates over the diff of the key range
@@ -190,4 +196,118 @@ func (i *MVCCIncrementalIterator) UnsafeKey() engine.MVCCKey {
 // the next call to {Next,Reset,Close}.
 func (i *MVCCIncrementalIterator) UnsafeValue() []byte {
 	return i.iter.UnsafeValue()
+}
+
+// RowCounter counts rows and index entires.
+type RowCounter struct {
+	prev roachpb.Key
+	roachpb.BulkOpSummary
+}
+
+// Count determines if the passed key is a new row and updates its counts if so.
+func (r *RowCounter) Count(key roachpb.Key) error {
+	// EnsureSafeSplitKey is usually used to avoid splitting a row across ranges,
+	// by returning the row's key prefix.
+	// We reuse it here to count "rows" by counting when it changes.
+	// Non-SQL keys are returned unchanged or may error -- we ignore them, since
+	// non-SQL keys are obviously thus not SQL rows.
+	row, err := keys.EnsureSafeSplitKey(key)
+	if err != nil || len(key) == len(row) {
+		return nil
+	}
+
+	// no change key prefix => no new row.
+	if bytes.Equal(row, r.prev) {
+		return nil
+	}
+	r.prev = append(r.prev[:0], row...)
+
+	rest, tbl, err := keys.DecodeTablePrefix(row)
+	if err != nil {
+		return err
+	}
+
+	if tbl < keys.MaxReservedDescID {
+		r.SystemRecords++
+	} else {
+		if _, indexID, err := encoding.DecodeUvarintAscending(rest); err != nil {
+			return err
+		} else if indexID == 1 {
+			r.Rows++
+		} else {
+			r.IndexEntries++
+		}
+	}
+
+	return nil
+}
+
+// MVCCIncIterExport exports changes to the keyrange [startKey,endKey) over the
+// interval (startTime, endTime]. Passing allRevisions exports every revision of
+// a key for the interval, otherwise only the latest value within the interval
+// is exported. Deletions are included if all revisions are requested or if the
+// startTime is non-zero. Returns the bytes of an SSTable containing the
+// exported keys, the number of keys exported and size of exported data, or an
+// error.
+func MVCCIncIterExport(
+	ctx context.Context,
+	e engine.Reader,
+	startKey, endKey roachpb.Key,
+	startTime, endTime hlc.Timestamp,
+	allRevisions bool,
+) ([]byte, roachpb.BulkOpSummary, error) {
+	var rows RowCounter
+
+	sst, err := engine.MakeRocksDBSstFileWriter()
+	if err != nil {
+		return nil, rows.BulkOpSummary, err
+	}
+	defer sst.Close()
+
+	// TODO(dan): Consider checking ctx periodically during the MVCCIterate call.
+	iter := NewMVCCIncrementalIterator(e, startTime, endTime)
+	defer iter.Close()
+	for iter.Seek(engine.MakeMVCCMetadataKey(startKey)); ; {
+		ok, err := iter.Valid()
+		if err != nil {
+			// The error may be a WriteIntentError. In which case, returning it will
+			// cause this command to be retried.
+			return nil, rows.BulkOpSummary, err
+		}
+		if !ok || iter.UnsafeKey().Key.Compare(endKey) >= 0 {
+			break
+		}
+
+		// Skip tombstone (len=0) records when startTime is zero
+		// (non-incremental) and we're not exporting all versions.
+		if !allRevisions && (startTime == hlc.Timestamp{}) && len(iter.UnsafeValue()) == 0 {
+			iter.NextKey()
+			continue
+		}
+
+		if log.V(3) {
+			v := roachpb.Value{RawBytes: iter.UnsafeValue()}
+			log.Infof(ctx, "Export %s %s", iter.UnsafeKey(), v.PrettyPrint())
+		}
+
+		if err := rows.Count(iter.UnsafeKey().Key); err != nil {
+			return nil, rows.BulkOpSummary, errors.Wrapf(err, "decoding %s", iter.UnsafeKey())
+		}
+		if err := sst.Add(engine.MVCCKeyValue{Key: iter.UnsafeKey(), Value: iter.UnsafeValue()}); err != nil {
+			return nil, rows.BulkOpSummary, errors.Wrapf(err, "adding key %s", iter.UnsafeKey())
+		}
+		if allRevisions {
+			iter.Next()
+		} else {
+			iter.NextKey()
+		}
+	}
+
+	rows.BulkOpSummary.DataSize = sst.DataSize
+	if sst.DataSize == 0 {
+		return nil, rows.BulkOpSummary, nil
+	}
+
+	sstContents, err := sst.Finish()
+	return sstContents, rows.BulkOpSummary, err
 }