Export involves iterating rocksdb (making a cgo call for every key iterated) and writing the results to an sstablewriter (another cgo call).
Based on some of the performance experiments on SSTable iteration using pure-Go LevelDB vs cgo'ed RocksDB, the expectation is that this will be considerably cheaper.