@petermattis 2017-11-29
> Based on some of the performance experiments on SSTable iteration using pure-Go LevelDB vs cgo'ed RocksDB, the expectation is that this will be considerably cheaper.

Shouldn't you be backing up that expectation with some benchmark numbers?

---

Review status: 0 of 13 files reviewed at latest revision, all discussions resolved.

---



*Comments from [Reviewable](https://reviewable.io:443/reviews/cockroachdb/cockroach/20328#-:-L-8-GLXcBGvzzKvldRT:b-yktu77)*
<!-- Sent from Reviewable.io -->