@benesch 2017-11-29
Haven't reviewed the meat of the change, yet, but build stuff looks good! Left you a few nits below.

---

Reviewed 13 of 13 files at r1.
Review status: all files reviewed at latest revision, 4 unresolved discussions, some commit checks pending.

---

*[c-deps/libroach/CMakeLists.txt, line 47 at r1](https://reviewable.io:443/reviews/cockroachdb/cockroach/20328#-L-8-tc0uooL25Bx9pGb:-L-8-tc0uooL25Bx9pGc:b-s2eoji) ([raw file](https://github.com/cockroachdb/cockroach/blob/c74ae2c9d5032176e6c2844f64cc95a71457ecc0/c-deps/libroach/CMakeLists.txt#L47)):*
> ```CMake
> )
> target_include_directories(roachccl
>   PRIVATE ../protobuf/src ../rocksdb/include protos
> ```

This is my bad, but we don't actually use the `protos` include directory because we specify relative paths to the proto headers. Can you instead remove it from both roach and roachccl?

---

*[c-deps/libroach/db.h, line 54 at r1](https://reviewable.io:443/reviews/cockroachdb/cockroach/20328#-L-80KfFrvS6YP2Z63V3:-L-80KfFrvS6YP2Z63V4:b-799esx) ([raw file](https://github.com/cockroachdb/cockroach/blob/c74ae2c9d5032176e6c2844f64cc95a71457ecc0/c-deps/libroach/db.h#L54)):*
> ```objc
> DBStatus DBSstFileWriterAddRaw(DBSstFileWriter* fw, const std::string& key, DBSlice val);
> 
> const DBStatus kSuccess = { NULL, 0 };
> ```

optional nit: I'd move this to the top of the file, before the func declarations.

---

*[pkg/ccl/storageccl/engineccl/rocksdb.go, line 175 at r1](https://reviewable.io:443/reviews/cockroachdb/cockroach/20328#-L-81enrOQYI7Mctah94:-L-81enrOQYI7Mctah95:b-x8vz6c) ([raw file](https://github.com/cockroachdb/cockroach/blob/c74ae2c9d5032176e6c2844f64cc95a71457ecc0/pkg/ccl/storageccl/engineccl/rocksdb.go#L175)):*
> ```Go
> 
> 	timeBoundIter := e.NewTimeBoundIterator(startTime, endTime)
> 	dbiter := (*C.DBIterator)(engine.GetRawIter(timeBoundIter))
> ```

This deserves a comment.

---

*[pkg/storage/engine/rocksdb.go, line 1913 at r1](https://reviewable.io:443/reviews/cockroachdb/cockroach/20328#-L-80hhjgJbNcNS2uIHM:-L-80hhjgJbNcNS2uIHN:b-7j8trw) ([raw file](https://github.com/cockroachdb/cockroach/blob/c74ae2c9d5032176e6c2844f64cc95a71457ecc0/pkg/storage/engine/rocksdb.go#L1913)):*
> ```Go
> 		return nil
> 	}
> ```

stray

---


*Comments from [Reviewable](https://reviewable.io:443/reviews/cockroachdb/cockroach/20328#-:-L-82OivSmMopHIjvu1m:bw266uu)*
<!-- Sent from Reviewable.io -->