@dt 2017-11-30
Yep, still needs more testing and benchmarking, just wanted to get some early feedback / make sure my approach wasn't way off while I fixed up our ccl benchmarks (they'd rotted a bit with local exportstorage changes and wouldn't run out of the box). I'm going to spin up a roachproad cluster to test the p99 impact today.

Added BenchmarkExport using the timebound iterator testdata but with 5 revisions per key instead of just one, on my mbp with a disk-backed rocksdb holding ~2gb of 64kb values, exporting roughly half of its key and time range with the go iterator vs cpp one:

```
name      old time/op    new time/op    delta
Export-8     631µs ± 7%     556µs ± 1%  -11.86%  (p=0.000 n=10+9)

name      old speed      new speed      delta
Export-8   564GB/s ± 7%   639GB/s ± 1%  +13.32%  (p=0.000 n=10+9)

name      old alloc/op   new alloc/op   delta
Export-8    4.59kB ± 0%    4.59kB ± 0%     ~     (all equal)

name      old allocs/op  new allocs/op  delta
Export-8      0.00           0.00          ~     (all equal)
```
 

---

Review status: 7 of 18 files reviewed at latest revision, 4 unresolved discussions.

---

*[c-deps/libroach/CMakeLists.txt, line 47 at r1](https://reviewable.io:443/reviews/cockroachdb/cockroach/20328#-L-8-tc0uooL25Bx9pGb:-L-87ZLhGnqpQTMDK9rA:b-ac6tcz) ([raw file](https://github.com/cockroachdb/cockroach/blob/c74ae2c9d5032176e6c2844f64cc95a71457ecc0/c-deps/libroach/CMakeLists.txt#L47)):*
<details><summary><i>Previously, benesch (Nikhil Benesch) wrote…</i></summary><blockquote>

This is my bad, but we don't actually use the `protos` include directory because we specify relative paths to the proto headers. Can you instead remove it from both roach and roachccl?
</blockquote></details>

I think the proto generated files themselves include e.g. `roachpb/metadata.pb.h` so they fail to compile if I remove it.

---

*[c-deps/libroach/db.h, line 54 at r1](https://reviewable.io:443/reviews/cockroachdb/cockroach/20328#-L-80KfFrvS6YP2Z63V3:-L-87HQMdCPKbFsoMJKY:b-896fix) ([raw file](https://github.com/cockroachdb/cockroach/blob/c74ae2c9d5032176e6c2844f64cc95a71457ecc0/c-deps/libroach/db.h#L54)):*
<details><summary><i>Previously, benesch (Nikhil Benesch) wrote…</i></summary><blockquote>

optional nit: I'd move this to the top of the file, before the func declarations.
</blockquote></details>

Done.

---

*[pkg/ccl/storageccl/engineccl/rocksdb.go, line 175 at r1](https://reviewable.io:443/reviews/cockroachdb/cockroach/20328#-L-81enrOQYI7Mctah94:-L-87CCyprvGGWf6Pc-0:b-896fix) ([raw file](https://github.com/cockroachdb/cockroach/blob/c74ae2c9d5032176e6c2844f64cc95a71457ecc0/pkg/ccl/storageccl/engineccl/rocksdb.go#L175)):*
<details><summary><i>Previously, benesch (Nikhil Benesch) wrote…</i></summary><blockquote>

This deserves a comment.
</blockquote></details>

Done.

---

*[pkg/storage/engine/rocksdb.go, line 1913 at r1](https://reviewable.io:443/reviews/cockroachdb/cockroach/20328#-L-80hhjgJbNcNS2uIHM:-L-85G4dJqHqGjJf9rCv:b-896fix) ([raw file](https://github.com/cockroachdb/cockroach/blob/c74ae2c9d5032176e6c2844f64cc95a71457ecc0/pkg/storage/engine/rocksdb.go#L1913)):*
<details><summary><i>Previously, benesch (Nikhil Benesch) wrote…</i></summary><blockquote>

stray
</blockquote></details>

Done.

---


*Comments from [Reviewable](https://reviewable.io:443/reviews/cockroachdb/cockroach/20328#-:-L-Cb4r2U3ED3y4V9-xY:boch1be)*
<!-- Sent from Reviewable.io -->