@benesch 2017-11-30
Reviewed 3 of 11 files at r2.
Review status: 10 of 18 files reviewed at latest revision, 3 unresolved discussions, some commit checks failed.

---

*[c-deps/libroach/CMakeLists.txt, line 47 at r1](https://reviewable.io:443/reviews/cockroachdb/cockroach/20328#-L-8-tc0uooL25Bx9pGb:-L-DfQia33lm0bNrOz1o:bdvfxa7) ([raw file](https://github.com/cockroachdb/cockroach/blob/c74ae2c9d5032176e6c2844f64cc95a71457ecc0/c-deps/libroach/CMakeLists.txt#L47)):*
<details><summary><i>Previously, dt (David Taylor) wrote…</i></summary><blockquote>

I think the proto generated files themselves include e.g. `roachpb/metadata.pb.h` so they fail to compile if I remove it.
</blockquote></details>

Blergh, ok. 

---

*[pkg/ccl/storageccl/engineccl/rocksdb.go, line 175 at r2](https://reviewable.io:443/reviews/cockroachdb/cockroach/20328#-L-Dfc2RNDZUIMgfb8BK:-L-Dfc2RNDZUIMgfb8BL:b-8s0ris) ([raw file](https://github.com/cockroachdb/cockroach/blob/ac325c4ee567201c0dffd7a73200f02139db38c2/pkg/ccl/storageccl/engineccl/rocksdb.go#L175)):*
> ```Go
> 
> 	timeBoundIter := e.NewTimeBoundIterator(startTime, endTime)
> 	// per-package gco stub generation means we need a cast, but it is just a
> ```

"gco"

---


*Comments from [Reviewable](https://reviewable.io:443/reviews/cockroachdb/cockroach/20328)*
<!-- Sent from Reviewable.io -->